Move templated class source into headers so they compile properly
Avoid C math functions
Actually implement get/set order in ParametersVirial
Construct integrators and invoke them from Zeno
